requirements.txt
django
requests
BeautifulSoup
google-generativeai
numpy


